refactor: 다이얼로그 함수 유틸화
